refactor(context): extract resolveComboContextLimit to its own module
contextManager.ts is under a frozen file-size ratchet (1001 lines) and the
combo-explicit branch pushed it to 1011, failing check:file-size.

Follow the gate's own guidance (modularize/extract rather than rebaseline):
move resolveComboContextLimit into open-sse/services/comboContextLimit.ts and
re-export it from contextManager so existing importers are unaffected.

contextManager.ts drops to 973 lines (below its frozen baseline) and the new
module documents the full precedence order in one place. No behavior change.
