+++ open-sse/services/contextManager.ts
@@ -436,50 +436,10 @@ export function resolveTokenLimit(
   return { limit: DEFAULT_LIMITS.default, specific: false };
 }
 
-/**
- * Resolve the context limit to use for proactive compression of a COMBO
- * request.
- *
- * chatCore always executes with the CONCRETE target's provider/model
- * (handleSingleModel resolves the target before delegating), so the
- * executing target's own limit is authoritative. Using min(...allTargets)
- * here — the previous behavior — compressed at the smallest sibling's
- * window even when running on the largest target, destructively purging
- * history long before the real window filled ("agent keeps forgetting").
- *
- * min(...comboTargetLimits) is kept only as a defensive fallback for the
- * case where the current provider/model resolves no specific limit at all.
- */
-export function resolveComboContextLimit(options: {
-  provider: string;
-  model: string | null;
-  comboTargetLimits: number[];
-  comboContextLength?: number | null;
-}): { limit: number; source: "combo-explicit" | "target" | "combo-min" | "fallback" } {
-  // An operator-set `context_length` on the combo record is an explicit
-  // declaration of the window this combo runs at, so it outranks every
-  // inferred value — including a target's own registry/catalog window.
-  // Without this branch the field was accepted and persisted by the combo
-  // editor but never consulted at request time, so a combo whose members
-  // carry no per-model window fell through to the provider's generic
-  // `defaultContextLength` and rejected large requests the operator had
-  // explicitly sized for.
-  const explicit = options.comboContextLength;
-  if (typeof explicit === "number" && Number.isFinite(explicit) && explicit > 0) {
-    return { limit: explicit, source: "combo-explicit" };
-  }
-  const own = resolveTokenLimit(options.provider, options.model ?? null);
-  if (own.specific) {
-    return { limit: own.limit, source: "target" };
-  }
-  const knownTargets = (options.comboTargetLimits || []).filter(
-    (value) => Number.isFinite(value) && value > 0
-  );
-  if (knownTargets.length > 0) {
-    return { limit: Math.min(...knownTargets), source: "combo-min" };
-  }
-  return { limit: own.limit, source: "fallback" };
-}
+// Combo context-limit resolution lives in ./comboContextLimit.ts; re-exported
+// here so existing importers keep working.
+export { resolveComboContextLimit } from "./comboContextLimit.ts";
+export type { ComboContextLimitSource } from "./comboContextLimit.ts";
 
 /**
  * Apply context compression to request body.
